+++ packages/headless/src/features/search-parameters/search-parameter-actions.ts
@@ -25,6 +25,7 @@ export type SearchParameters = Omit<
   | 'searchHub'
   | 'pipeline'
   | 'context'
+  | 'locale'
 > &
   Partial<FacetParameters>;