+++ packages/atomic/CHANGELOG.md
@@ -3,6 +3,83 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.7.0-alpha.24](https://github.com/coveo/ui-kit/compare/v0.7.0-alpha.23...v0.7.0-alpha.24) (2021-02-26)
+
+**Note:** Version bump only for package @coveo/atomic
+
+
+
+
+
+# [0.7.0-alpha.23](https://github.com/coveo/ui-kit/compare/v0.7.0-alpha.22...v0.7.0-alpha.23) (2021-02-26)
+
+
+### Features
+
+* **atomic:** facet Sort Criterion and Enable Facet Search Prop ([#543](https://github.com/coveo/ui-kit/issues/543)) ([a2c50c3](https://github.com/coveo/ui-kit/commit/a2c50c357144a024a7dca551ada59f4e72c6f9ff))
+
+
+
+
+
+# [0.7.0-alpha.22](https://github.com/coveo/ui-kit/compare/v0.7.0-alpha.21...v0.7.0-alpha.22) (2021-02-26)
+
+
+### Features
+
+* **atomic:** category Facet ([#542](https://github.com/coveo/ui-kit/issues/542)) ([cdd2c01](https://github.com/coveo/ui-kit/commit/cdd2c01c0eb34585c703f03fd6b764723c77b3e0))
+
+
+
+
+
+# [0.7.0-alpha.21](https://github.com/coveo/ui-kit/compare/v0.7.0-alpha.20...v0.7.0-alpha.21) (2021-02-25)
+
+
+### Bug Fixes
+
+* **atomic:** fix search box styles ([#546](https://github.com/coveo/ui-kit/issues/546)) ([13a85aa](https://github.com/coveo/ui-kit/commit/13a85aaad37d737f0a7d9b8e3fc9e0533742436b))
+
+
+### Features
+
+* **atomic:** improve styling solution for page layout & templates ([#538](https://github.com/coveo/ui-kit/issues/538)) ([afb53b4](https://github.com/coveo/ui-kit/commit/afb53b426d217e8a10a44b20ab5d090a7c350b9c))
+
+
+
+
+
+# [0.7.0-alpha.20](https://github.com/coveo/ui-kit/compare/v0.7.0-alpha.19...v0.7.0-alpha.20) (2021-02-24)
+
+**Note:** Version bump only for package @coveo/atomic
+
+
+
+
+
+# [0.7.0-alpha.19](https://github.com/coveo/ui-kit/compare/v0.7.0-alpha.18...v0.7.0-alpha.19) (2021-02-24)
+
+
+### Features
+
+* **atomic:** date facet component  ([#534](https://github.com/coveo/ui-kit/issues/534)) ([f6a8cdd](https://github.com/coveo/ui-kit/commit/f6a8cddf5d4820b3ccdea393f93f4317df1a0e7c))
+* **atomic:** facet component ([#535](https://github.com/coveo/ui-kit/issues/535)) ([a186c5a](https://github.com/coveo/ui-kit/commit/a186c5a251c8915533cb0d6374be4ac7cf04b42c))
+
+
+
+
+
+# [0.7.0-alpha.17](https://github.com/coveo/ui-kit/compare/v0.7.0-alpha.16...v0.7.0-alpha.17) (2021-02-23)
+
+
+### Features
+
+* **atomic:** numeric facet ([#521](https://github.com/coveo/ui-kit/issues/521)) ([2259add](https://github.com/coveo/ui-kit/commit/2259add97432eb6ce6c57c53078aa6658a228389))
+
+
+
+
+
 # [0.7.0-alpha.16](https://github.com/coveo/ui-kit/compare/v0.7.0-alpha.15...v0.7.0-alpha.16) (2021-02-23)
 
 **Note:** Version bump only for package @coveo/atomic